producers/reddit/requirements.txt
kafka-python==2.0.2
praw==7.7.1
python-dotenv==1.0.0
colorlog==6.8.0
